Clarify block/page terminology
I was confused by the documentation of `DFUMemIO` because it seems to use "block" and "page" interchangeably. I wanted to clarify the terminology used and settled on the following:

A "block" is the largest buffer that can be read or programmed. A "page" is the smallest region of flash memory that can be erased.